+++ frontend/src/lib/api/ckbtc-ledger.api.ts
@@ -1,20 +1,13 @@
-import { createAgent } from "$lib/api/agent.api";
 import {
   getIcrcAccount,
   getIcrcToken,
-  icrcTransfer as transferIcrcApi,
-  type IcrcTransferParams,
+  icrcLedgerCanister,
 } from "$lib/api/icrc-ledger.api";
-import { HOST } from "$lib/constants/environment.constants";
 import type { Account, AccountType } from "$lib/types/account";
 import type { IcrcTokenMetadata } from "$lib/types/icrc";
 import { logWithTimestamp } from "$lib/utils/dev.utils";
-import type { HttpAgent, Identity } from "@dfinity/agent";
-import {
-  IcrcLedgerCanister,
-  type IcrcAccount,
-  type IcrcBlockIndex,
-} from "@dfinity/ledger";
+import type { Identity } from "@dfinity/agent";
+import type { IcrcAccount } from "@dfinity/ledger";
 import type { Principal } from "@dfinity/principal";
 
 export const getCkBTCAccount = async ({
@@ -32,7 +25,7 @@ export const getCkBTCAccount = async ({
 
   const {
     canister: { balance },
-  } = await ckBTCLedgerCanister({ identity, canisterId });
+  } = await icrcLedgerCanister({ identity, canisterId });
 
   const callParams = {
     certified,
@@ -62,7 +55,7 @@ export const getCkBTCToken = async ({
 
   const {
     canister: { metadata: getMetadata },
-  } = await ckBTCLedgerCanister({ identity, canisterId });
+  } = await icrcLedgerCanister({ identity, canisterId });
 
   const token = await getIcrcToken({
     certified,
@@ -72,54 +65,4 @@ export const getCkBTCToken = async ({
   logWithTimestamp("Getting ckBTC token: done");
 
   return token;
-};
-
-export const ckBTCTransfer = async ({
-  identity,
-  canisterId,
-  ...rest
-}: {
-  identity: Identity;
-  canisterId: Principal;
-} & Omit<IcrcTransferParams, "transfer">): Promise<IcrcBlockIndex> => {
-  logWithTimestamp("Getting ckBTC transfer: call...");
-
-  const {
-    canister: { transfer: transferApi },
-  } = await ckBTCLedgerCanister({ identity, canisterId });
-
-  const blockIndex = await transferIcrcApi({
-    ...rest,
-    transfer: transferApi,
-  });
-
-  logWithTimestamp("Getting ckBTC transfer: done");
-
-  return blockIndex;
-};
-
-const ckBTCLedgerCanister = async ({
-  identity,
-  canisterId,
-}: {
-  identity: Identity;
-  canisterId: Principal;
-}): Promise<{
-  canister: IcrcLedgerCanister;
-  agent: HttpAgent;
-}> => {
-  const agent = await createAgent({
-    identity,
-    host: HOST,
-  });
-
-  const canister = IcrcLedgerCanister.create({
-    agent,
-    canisterId,
-  });
-
-  return {
-    canister,
-    agent,
-  };
 };
