+++ frontend/src/tests/lib/api/icrc-ledger.api.spec.ts
@@ -1,15 +1,29 @@
 import {
+  executeIcrcTransfer,
   getIcrcAccount,
   getIcrcToken,
   icrcTransfer,
 } from "$lib/api/icrc-ledger.api";
+import { CKBTC_LEDGER_CANISTER_ID } from "$lib/constants/ckbtc-canister-ids.constants";
 import { mockIdentity } from "$tests/mocks/auth.store.mock";
 import {
   mockQueryTokenResponse,
   mockSnsToken,
 } from "$tests/mocks/sns-projects.mock";
+import { IcrcLedgerCanister } from "@dfinity/ledger";
+import mock from "jest-mock-extended/lib/Mock";
 
 describe("icrc-ledger api", () => {
+  const ledgerCanisterMock = mock<IcrcLedgerCanister>();
+
+  beforeEach(() => {
+    jest
+      .spyOn(IcrcLedgerCanister, "create")
+      .mockImplementation(() => ledgerCanisterMock);
+  });
+
+  afterEach(() => jest.clearAllMocks());
+
   describe("getIcrcMainAccount", () => {
     it("returns main account with balance and project token metadata", async () => {
       const balanceSpy = jest.fn().mockResolvedValue(BigInt(10_000_000));
@@ -79,16 +93,34 @@ describe("icrc-ledger api", () => {
     });
   });
 
-  describe("transfer", () => {
+  describe("execute transfer", () => {
     it("successfully calls transfer api", async () => {
       const transferSpy = jest.fn().mockResolvedValue(undefined);
 
-      await icrcTransfer({
+      await executeIcrcTransfer({
         to: { owner: mockIdentity.getPrincipal() },
         amount: BigInt(10_000_000),
         createdAt: BigInt(123456),
         fee: BigInt(10_000),
         transfer: transferSpy,
+      });
+
+      expect(transferSpy).toBeCalled();
+    });
+  });
+
+  describe("transfer", () => {
+    it("successfully calls transfer api", async () => {
+      const transferSpy =
+        ledgerCanisterMock.transfer.mockResolvedValue(undefined);
+
+      await icrcTransfer({
+        identity: mockIdentity,
+        to: { owner: mockIdentity.getPrincipal() },
+        amount: BigInt(10_000_000),
+        createdAt: BigInt(123456),
+        fee: BigInt(10_000),
+        canisterId: CKBTC_LEDGER_CANISTER_ID,
       });
 
       expect(transferSpy).toBeCalled();
